Moving obstacle; Code cleanup; BG Music/Collision sound; Game ends at 0 lives
