+++ Assets/Prefab/Capsule.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 49d8acdc3817376439278eeeb02e21c7
+timeCreated: 1491607008
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
