+++ Assets/Prefab/Obstacle/InvTrigger.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: dfee83fc11460164a872bc6933329436
+timeCreated: 1492212379
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
