+++ Assets/Prefab/Tracks/TrackMoving.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8466f8b18be72374fb434b7acac790b1
+timeCreated: 1494023286
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
