--- Assets/Prefabs.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 24edc90ede9e9d34aa188e5790b3e1cd
-folderAsset: yes
-timeCreated: 1492212374
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
