--- Assets/Prefabs/InvTrigger.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: dfee83fc11460164a872bc6933329436
-timeCreated: 1492212379
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
