--- Assets/Scripts/GameManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: e67c1addf25b212469bd56329cdcef04
-timeCreated: 1492832126
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
