+++ Assets/Scripts/ObstacleMove.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5121722b271083c4996917e0c3aa80bc
+timeCreated: 1494022322
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
