+++ Assets/Scripts/PlayerScripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a630013ea979a8b4f810f116032c66dc
+folderAsset: yes
+timeCreated: 1494029961
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
