+++ Assets/Scripts/PlayerScripts/AccController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 77262ef0e834f594f86a2aa06cbdf145
+timeCreated: 1491608579
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
