+++ Assets/Scripts/PlayerScripts/DisplayText.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a07e5570ea8dda249b86e5974e568f43
+timeCreated: 1494030018
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
