+++ Assets/Scripts/PlayerScripts/ScoreManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: df4377def4c045543b882afa4e848628
+timeCreated: 1491174076
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
