+++ Assets/Scripts/PlayerScripts/Track.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f6edccf55243fe84f9585ed9c6579c42
+timeCreated: 1491608579
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
