+++ Assets/Scripts/PlayerScripts/TrackGeneration.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 15f471324b0ee774780908235dc57b7e
+timeCreated: 1492223091
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
