+++ Assets/Sound.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 0d7efc98da8e9a047adba4b9b8282c6c
+folderAsset: yes
+timeCreated: 1494026301
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
