+++ Assets/test.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bcf5eee9f2ad1e04bb40db1244e8a5b4
+timeCreated: 1494022656
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
