+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 5.5.1f1
+m_EditorVersion: 5.6.0f3
